add ArgsParser, tilemap_editor executable

+++ Rendering/TextureAtlas.cpp
@@ -11,9 +11,9 @@ TextureAtlas::TextureAtlas(Managed<sf::Texture> atlas, size_t subtexture_size)
     }
 }
 
-std::pair<vec<float>, vec<float> > TextureAtlas::subtexture_coords(vec<size_t> subtexture_index) const {
+std::pair<vec<float>, vec<float>> TextureAtlas::subtexture_coords(vec<size_t> subtexture_index) const {
     return {
-        vec<float>(subtexture_index.x * m_subtexture_size, subtexture_index.y * m_subtexture_size),
-        vec<float>((subtexture_index.x + 1) * m_subtexture_size, (subtexture_index.y + 1) * m_subtexture_size)
+        vec<float>(float(subtexture_index.x * m_subtexture_size), float(subtexture_index.y * m_subtexture_size)),
+        vec<float>(float((subtexture_index.x + 1) * m_subtexture_size), float((subtexture_index.y + 1) * m_subtexture_size))
     };
 }
